+++ src/lib/public_api.ts
@@ -42,6 +42,8 @@ export { StFilterList } from './pipes/search-filter.pipe';
 
 /**************************** DIRECTIVES ***************************************************/
 export { StInputAdjustable } from './directives/st-input-adjustable';
+export { StMinValidator } from './directives/form/st-min-validator/st-min-validator';
+export { StMaxValidator } from './directives/form/st-max-validator/st-max-validator';
 
 /**************************** MODULES *************************************************/